rendernode: Stop preregistering of GTypes
This is to simplify a future splitup of node types into separate .c
files.

This code could be slower but I wasn't able to measure the difference,
because surrouding code was causing so much fluctuation.
So it's hopefully not a big problem

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9125>