gsk: Add gdk_color_is_srgb()
Small utility function to check if a color fits into the sRGB color
volume.
It's essentially a check if we need to enable HDR support due to a
color.

For now, this just checks the colorstate, but in the future, when we
have color volumes, it will check the color is in the sRGB color volume.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9130>