vulkan: Rework swapchain present implementation
I had for some inexplicable reason assumed that the swapchain uses the
same queue as we submit on and therefor no synchronization is necessary.
That's not how it works though.
Well, it was technically true with most drivers, so nothing bad happened.

Until Mesa decided to make things run faster by using different queues.
Then all hell broke loose and GTK exploded left and right.

The actual implementation of proper synchronization was originally very
underspecified and was only properly specified with the
VK_EXT_swapchain_maintenance1 extension.
See
https://www.khronos.org/blog/resolving-longstanding-issues-with-wsi
for a discussion of the problems.

That's why we need 2 different codepaths: One when running with the
extension that uses fences to track lifetimes of objects and one that
uses best practices when fences aren't available and hopes for the best.

During this refactoring, I also untied the swapchain from the context,
so that GTK can now recreate the swapchain while there are still
outstanding images - we just keep the old swapchain alive. Now that we
do proper tracking, that's not too hard.

Unfortunately all of this is rather entangled, so I didn't manage to
split it into multiple commits.

This also fixes all previous validation layer complaints when running
with both
  VK_INSTANCE_LAYERS=VK_LAYER_KHRONOS_validation
  VK_LAYERS_ENABLES=VK_VALIDATION_FEATURE_ENABLE_SYNCHRONIZATION_VALIDATION_EXT

Fixes: https://gitlab.gnome.org/GNOME/gtk/-/issues/7866
Fixes: https://gitlab.freedesktop.org/mesa/mesa/-/issues/14087
Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9128>