paste: Ensure porter-duff operations don't pass through
Pasting is meant to work like a texture, so replaying it as nodes needs
to ensure that potential Porter/Duff operations of those nodes do not
end up affecting the background.

Implement this by wrapping the node in a 100% opaque opacity node,
because there is no explicit node for this purpose.

Test included.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9159>