gsk: make gsk_render_node_alloc() take a GType
It's just a wrapper around g_type_create_instance, so we can pass the
GType.

I could have removed it completely, but decided to keep the function so
that there's an easy place to grep for node creation and so that we have
a place where we could add initialization code should we ever want
to.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9125>