gpu: Implement support for GskCompositeNode
This support requires dual source blending for now.

It is implemented using a new shader that uses those dual source
blending features to compute the right masks for destination and source
and then relying on the new blend modes to produce the correct output.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9159>