testsuite: Add an "opaque" run to compare-render
This run gets the opaque region, shrinks it to the nearest integer, and
then checks that rendering the node clipped to that rectangle produces
an opaque result.

We skip all the nodes that do not report an opaque rectangle.

This tests both that the opaque rect computation is not to greedy (or
flat out wrong) and that the regions that should be opaque indeed are.

My local results:
                   GL      Vulkan  Cairo
Ok:                100     100     89
Fail:              0       0       0
Skipped:           129     129     115

So it tests about 43% of the compare tests, which is quite a nice
amount.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9160>