node editor: Unset GtkApplication:save-state
Currently, every time you close and restart the node editor,
it acts as if it crashed and refuses to restore state, which
defeats the purpose. Before we can enable the session saving
support, the preexisting functionality needs to be properly
integrated with the it.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9161>