application: Reduce warning spam
It is a bit annoying if we spam warnings just because you run
a recent GTK app on an old OS. Lower the warnings about lack
of session saving infrastructure to debug messages. We still
warn if we get an unexpected D-Bus error.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9161>