testsuite: Parse expected opaque region from file contents
Instead of looking at the filename, look for a comment in the first line
of the node file that is a rectangle, like so:

/* 0 25 100 50 */

If it exists, that is the expected opaque region. If it doesn't exist
the node is expected to not have an empty region.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9172>