css: Fix opacity handling of backdrop-filter
We need to take into account opacity.

I decided to use 2 opacity nodes - one for the backdrop filter, one for
the content, so that the copy node has as few children as possible, so
just the backdrop-filter itself.
And because the copy needs to happen before the opacity, there's no way
to have that opacity node apply to the whole content.

Test included

Fixes: #7880
Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9172>