gsk: Add GskRenderNode.copy_mode
It's an internal flag for optimizations that tracks 3 possible values:

* This node contains a copy node that will paste its contents

* Same as above, but the paste will not displace any pixels
  This is mainly relevant for clipping, because clipping does not need
  to preserve any pixels in this case.

* This node does not contain a copy node.

Note that this flag is different from "this node will be copied" because
copies happen in drawing order, not in parent/child order. Example:
  container {
    color {}
    copy {
      ...
    }
  }
The container will be marked by this flag (and so will the copy node),
but the color node will not. It is the job of the container node to
track that the color node will be copied.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9187>