offload: Handle multiple nodes using the same subsurface
This can happen in cases where nodes are used twice in the node tree.
The most common case is with a GtkWidgetPaintable.
In that case, a subsurface node can be duplicated into 2 places because
it's just part of the widget's rendernode.

Another place that makes this more visible is the recent copy/paste node
introduction which copies part of the node tree into the paste node's
place. And if that part contains subsurface nodes...

We now just don't offload in that case, which makes sense because we
could only offload once and the other location still needs to read the
texture.
And in that case it's cheaper to just not offload anywhere.

Note that this requires walking the whole node tree to determine if
there's another subsurface node anywhere and we can't just skip nodes
without subsurface nodes.
Which is why ported to GskRenderReplay while I was at it.

Fixes: #7885
Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9187>