filechooser: Fix filters being removed when navigating via the bar path
When navigating via the bar path, the filechooser try to select the child
within the path of the seleted folder.

The filter is disabled if any of the selected files (in our case only one)
is not visible.

The test for visibility was testing for filechooser::filtered-out property,
but it is filechooser::visible that makes an entry visible or not (see
match_func() in gtkfilechooserwidget.c).

An entry can be filtered and still visible, as in our case: folders.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9184>