gsk: Add GskRepeatNode::repeat
This will be necessary for drawing css-conformant backdrop-filter for
blurs, but it's been a feature we've been discussing before.

It's a bit useless to add GSK_REPEAT_NONE, but hey, consistency wins.

For now, that new API is private because it's waiting for snapping APIs
to land before making it public.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9200>