snapshot Add (private) gtk_snapshot_push_repeat2()
Pushes a repeat node with a GskRepeat. We need a better name if we want
to make this public API.

We could also decide to add gtk_snapshot_push_reflect() and
gtk_snapshot_push_pad() instead of adding the enum maybe?

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9200>