gsk: Add a new way to compute opacity
The new way creates a "renderer" object and has every node "render" its
opacity to it.
The vfunc also carries information about copy nodes, so that paste nodes
can know if they are opaque or not.

The old vfuncs still exist, and a compile-time switch exists in
gsk_render_node_get_opaque() to switch between the old and new version.