testsuite: Add a complex test for PORTER_DUFF_SOURCE
Draw a complex mask with PORTER_DUFF_SOURCE with the caveat that both the
source and the destination are opaque across the bounds of the mask.

Expect this to result in an opaque result.

The reason for this to be explicitly tested is that certain backdrop
operations for regions with rounded corners may cause this to happen.
Here's a simplified example:

  /* opaque stuff drawn here, like window background */
  copy {
    composite {
      child: blur {
        child: paste { ... }
        blur: 5;
      }
      mask: rounded-clip { ... }
      operator: source;
    }
  }

And it that case, we don't want to regress on the opaque region we
report to the compsitor.