gsk: Implement opacity for blur
A common operation is to blur the background and clip the result in such
a way that the result remains opaque.

Support this, so that a PORTER_DUFF_SOURCE operation can replace the
blurred background without a loss of opaque region.

Testcase demonstrating it included.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9210>