offload: Handle composite nodes compositing with the subsurface
When a CompositeNode overlaps a subsurface node, we cannot offload it.
The CompositeNode might modify the rendering in an unexpected way.

Also add 2 tests:
Check that a composite node before an offload does not influence the
result.
Check that a composite node after an offload stops that offload from
happening.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9226>