nodeparser: Don't name nested "duplicates"
When a node is referenced twice in the same node file, we want to give
it a name.

However, previously we would also consider all the children of this node
to be duplicated in the file, even though they were all only referenced
once.

So ensure that we don't name those.

Test included.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9219>