svg: Use the viewBox transformation specified in the spec
The math behind the viewbox transformation is specified here:
https://www.w3.org/TR/SVG2/coords.html#ComputingAViewportsTransform
and matches our compute_viewport_transform.

But at the end it says "The transform applied to content contained by
the element is given by
translate(translate-x, translate-y) scale(scale-x, scale-y)."

We previously applied the individual transform operations the wrong way
arround, but with this patch the order now follows the spec.

It just happen to work for all our current test cases, as they all use
a viewbox whoose top left corner is (0|0), so the tranlation
transformation was just the identity transform.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9220>