offload: Track clears-background and copy mode for containers
Subsurfaces whose contents get copied or which might be overdrawn by
composite nodes need to be tracked.

However, they will not necessarily be marked like this themselves when
they are put into a container in front of the copy or composite node.
So the container node needs to track this and disable offloading for
branches of the node tree where this is the case.

Some tests included

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9226>