renderer: Move copy/paste replacement out of generic code path
.. and into the renderers.

For now, all renderers do the copy paste replacement, but that can now
change.

As a side effect, the opaque node computations for the opaque rect now
use copy/paste nodes which don't yet properly track opacity.
That work is ongoing in !9210 but should not be noticeable in practice
(famous last words...)

And another side effect is that it makes copy/paste nodes work with
gsk_render_node_draw() which was overlooked in previous commits.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9230>