svg: Implement mix-blend-mode
This adds a mix-blend-mode attribute (that can only be set
via the CSS style property), and then implements it using
copy and blend nodes.

To make the backdrop copy not copy too much, we wrap the
entire rendering into an opacity node, which has the effect
of 'isolating' the group, as CSS calls it.

All renderer tests that are affected by the isolation have
been updated.

Test included.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9232>