svg: Introduce css-only attributes
SVG allows to just set any CSS property in style attributes.

If we want to handle some more of that, we need a way to store
these attributes without allowing random nonsense that no other
SVG parser accepts.

This first candidate to use this for is mix-blend-mode.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9232>