inspector: Make copy/paste nodes display properly
Run gsk_render_node_replace_copy_paste() on the root node and generate a
"draw node" that is used for display.

That node contains the right replaced contents for the original node, so
the visuals appear as expected.

It's a bit finicky because the 2 nodes need to be kept in sync, but I
hope I got it right.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9233>