API: Add GtkPicture::isolate-contents
This allows turning off the isolation that we enforce on paintables in
the few cases where applications want them to meddle with the CSS-drawn
widgettry.

The isolation properties that are available in GSK are more fine grained
but after discussion it was determined that it's hard to provide a
forward-compatible API that does the right thing when new isolation
options are added, yet provides an easy to use property to application
developers.

So this widget only provides a boolean property: Either the default
isolation (TRUE) or no isolation at all (FALSE). If people want
fine-grained isolation, they can set the property to FALSE and wrap the
widget in custom code that calls gtk_snapshot_push_isolation() with
exactly the features they want.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9235>