svg: Add a content fit attribute
Add a SvgContentFit value type, and use it for parsing the
preserveAspectRatio attribute on the svg element. Also add
an attribute with values of that type. So far, no shape is
using it, but it will be used by patterns in the future.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9249>