a11y: Fix the accessible range max value
According to the ARIA documentation [^1] and [^2], `valuenow` should
be between `valuemin` and `valuemax` inclusive, as it may be
presented as a percentage.

In gtk, it means the page size shall be subtracted from the upper
limit when using a GtkAdjustment.

This matches what was done back then in gtk+3 [^3], and also seems
compliant with Qt implementation [^4].

[^1]: https://www.w3.org/TR/wai-aria-1.2/#scrollbar
[^2]: https://www.w3.org/TR/wai-aria-1.2/#scrollbar
[^3]: https://gitlab.gnome.org/GNOME/gtk/-/blob/3.24.50/gtk/a11y/gtkrangeaccessible.c#L141
[^4]: https://doc.qt.io/qt-6/qscrollbar.html

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9261>