aspectframe: Reimplement measurements
It is incorrect to just pass through the measurements of the child.

The natural size we return should be proportional to for-size, so the
ancestor widgets are encouraged to give us an allocation that already
has the desired aspect ratio; this is similar to what GtkPicture does.

For the minimum size, consider that we always give our child an
allocation that has the desired aspect ratio. Among such allocations,
there is a single smallest one that is still acceptable to the child
according to its size request. Aspect frame's allocation can then be as
small as this minimum allocation, but no smaller. Moreover, aspect
frame's minimum size in one orientation therefore doesn't depend on the
available size in the other orientation, since it always equals to the
size of that minimum child allocation.

To find this minimum allocation size, we have to use binary searching:
among the sizes (width x height pairs) that are acceptable to the child,
we're looking for the one that fits our desired ratio best. This is not
unlike what gtk_window_compute_min_size () does.

Since this binary searching is expensive and does not depend on
for-size, it is desirable to cache its outcomes between calls to
measure (). Note that while the minimum size of an aspect frame does not
depend on for-size, the natural size does, so we cannot report ourselves
as constant-size. Usually, SizeRequestCache / _gtk_size_request_cache
helps with avoiding measure vfunc calls when the widget's measurements
stay constant over a range of for-size values. That doesn't apply here
however, since our reported natural sizes still differ with for-size
(even though the minimum sizes don't), therefore SizeRequestCache won't
short-curcuit measurement requests, and we'd be repeating the binary
search over and over, for every for-size.

So instead, we make use of the internal resize_func mechanism and
implement caching of the minimum size on the widget side. resize_func is
invoked whenever the widget is resized (as in queue_resize), and we
clear our cache there. Between resizes, only the first call to the
measure vfunc will cause the binary search to run, and any later ones
will pick up the cached minimum size -- while still computing the
natural size proportionally to the passed in for-size every time.

There is another complication because of fractional values: in many
cases, we're dealing with sizes whose aspect ratio is close enough to
the desired one, but not exactly equal to it (which may not even be
achivable, or only achievable with huge width & height values). Exactly
what constitutes an acceptable allocation is not very clearly defined,
and trying to define it more precisely (and actually follow that
definition in the implementation), while desirable, seems to bring more
issues than it solves. In other words, the "close enough" approach as
implemented in this patch appears to be most pragmatic; but it leads to
the natural size, computed proportionally to the size in the opposite
orientation, potentially being smaller than the minimum size, which
wouldn't happen if acceptable allocations were defined more precisely.
As a workaround that works good enough, we simply MAX the computed
natural size with the minimum size.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>
Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9264>