svg: Stop allocating too many values
Make the value arrays in Shape only contain actual shape
attributes. This decouples the size of the Shape struct
from the growth of the attribute enum, which is a good
thing, since filter are going to require a bunch more
attributes. And they are only ever used on filters.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9274>