copypasteutils: Return empty container instead of NULL
Doing this guarantees that the number of nodes doesn't shrink when
copy.paste nodes are replaced, which is necessary for the inspector
tracking 2 node trees - the actual one and the replaced one for
rendering paste nodes properly.

Fixes assertions triggering when a copy nodes copied nothing.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9279>