imcontextwayland: Store information about pre-edit style hints
This is at the moment unused, only stored and freed in the places it
would be expected. The information is stored as an ordered array of
non-contiguous segments, along with a set of flags for the style
hints set for the segment.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/6922>