imcontextwayland: Handle conversion of pre-edit style hints to pango
This is hooked up to the CSS machinery, so that a "preedit" node with
a set of style classes for the (combined) hints is used. These CSS nodes
are cached by the Wayland GtkIMContext implementation as long as the
object has a client widget. Optionally it can be made to use a different
parent CSS node than the widget's own.

When obtaining the pre-edit text, these CSS nodes will be looked up
and converted to a PangoAttrList ready to use for the widget.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/6922>