svg: Sort out accumulation for transforms
The small print in the SVG animation spec reveals
why the browsers behave so myseriously if a mix
of <animateTransform> and <animateMotion> applies
to a shape: they are meant to accumulate separately,
with the accumulated motion always applying on
top of the accumulated transform.

This makes our behavior match the browsers for
the animate-elem-24-t.svg testcase in the svg 1.1
testsuite.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9297>