gpu: Fix pad color for gradients
If multiple color stops exist with offset 0 or 1, ensure that
GSK_REPEAT_PAD picks the first or last color stop of the gradient and
not any other one with offset 0 or 1.

I encountered this while playing around in node-editor.

Tests included.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9352>