svg: Allow embedding fonts
Repurpose the SVG 1.1 font-face element and allow
font-face-uri to carry a data url with a base64
encoded font file. The main target for this
is the testsuite, where we want to have testcases
that are independent of the system font setup.

This is strongly based on the way we handle
embedded fonts in node files.

All the renderer tests have been converted to
embed the fonts that they use.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9309>