svg: limit rendering up to an upper bound
To prevent SVGs from causing a denial-of-service, limit the maximum
amount of elements the SVG renderer can parse and consume.

This was inspired by similar protection measures done by rsvg, however
as we rely on the GPU for rendering instead of cairo, the bounds have
been significantly decreased, as cairo is *a lot* better at dealing with
many drawcalls.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9310>