build: Accept -I arguments when processing glsl headers
I want to autogenerate some of the GLSL code and that code of course
goes into the build directory, so we need to look there, too.

Also allow passing multiple files for concatenation. I experimented with
this first, but it didn't work - still, why not keep it?

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9321>