gpu: Redo vertex argument passing
In particular, rewrite generate-header.py into a more flexible tool.

The tool reads a preamble in the GLSL and generates the known header
file and a GLSL include file that declares the vertex shader inputs.

As a bonus, it collects multiple smaller arguments into a vec4 to save
space.

For now, the initial version of the tool exists and the boxshadow shader
has been converted as a proof of concept.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9321>