svg: Implement non-scaling strokes
This amounts to placing the stroke node in host coordinates,
while keeping the child providing the paint in user coordinates.
Placing the stroke node in host coordinates means that
we need to transform the path to compensate.

The node structure is as follows:

  transform to-host
    stroke transformed-path
      transform to-user
        child

Note that this is currently implemented for stroking shapes,
but not for stroking text.

Test included.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9319>